requirements.txt
streamlit>=1.33
supabase>=2.3
python-jose[cryptography]
praw
python-dotenv>=1.0.0
openai
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
PyJWT>=2.8.0
resend>=0.6.0
